+++ frontend/src/app/calendar/[id]/page.jsx
@@ -1,0 +1,5 @@
+const CalenderDetail = () => {
+    // TODO: 일정 보기/수정/삭제 페이지 구현
+}
+
+export default CalenderDetail;
