+++ frontend/src/app/calendar/new/page.jsx
@@ -1,0 +1,5 @@
+const CalenderNew = () => {
+    // TODO: 일정 추가 페이지 구현
+}
+
+export default CalenderNew;
