+++ frontend/src/app/calendar/page.jsx
@@ -1,0 +1,5 @@
+const CalendarMain = () => {
+    // TODO: 캘린더 페이지 구현
+}
+
+export default CalendarMain;
