+++ frontend/src/app/layout.jsx
@@ -1,0 +1,26 @@
+import localFont from "next/font/local";
+import "@/styles/globals.css";
+
+const wantedSans = localFont({
+    src: "../fonts/WantedSansVariable.woff2",
+    preload: true,
+    style: "normal",
+    display: "swap",
+    weight: "400 1000",
+    variable: "--font-wanted-sans",
+});
+
+export const metadata = {
+    title: "Managers",
+    description: "일정 관리 프로젝트",
+};
+
+export default function RootLayout({ children }) {
+    return (
+        <html lang="ko">
+            <body className={`${wantedSans.variable}`}>
+                {children}
+            </body>
+        </html>
+    );
+}
