+++ frontend/src/app/login/page.jsx
@@ -1,0 +1,5 @@
+const Login = () => {
+    // TODO: 로그인 페이지 구현
+}
+
+export default Login;
