+++ frontend/src/app/register/page.jsx
@@ -1,0 +1,5 @@
+const Register = () => {
+    // TODO: 회원가입 페이지 구현
+}
+
+export default Register;
