+++ frontend/src/app/user/page.jsx
@@ -1,0 +1,5 @@
+const User = () => {
+    // TODO: 유저 설정 페이지(비밀번호 변경 등) 구현
+}
+
+export default User;
